// src/openapi3/aether/2.0.0/models/site.ts
// GENERATED CODE -- DO NOT EDIT!
/* tslint:disable */
/* eslint-disable */
import { AdditionalPropertyTarget } from './additional-property-target';
import {EnterpriseEnterpriseSite} from "./enterprise-enterprise-site";

/**
 * The top level container
 */
export interface Site {

  /**
   * List of site
   */
  site?: Array<EnterpriseEnterpriseSite>;

  [key: string]: AdditionalPropertyTarget | Array<EnterpriseEnterpriseSite> | undefined;
}