// src/openapi3/aether/2.0.0/models/upf.ts
// GENERATED CODE -- DO NOT EDIT!
/* tslint:disable */
/* eslint-disable */
import { AdditionalPropertyTarget } from './additional-property-target';
import {EnterpriseEnterpriseSiteUpf} from "./enterprise-enterprise-site-upf";

/**
 * The top level container
 */
export interface Upf {

  /**
   * A list of named upfs.
   */
  upf?: Array<EnterpriseEnterpriseSiteUpf>;

  [key: string]: AdditionalPropertyTarget | Array<EnterpriseEnterpriseSiteUpf> | undefined;
}