// src/openapi3/aether/2.0.0/models/vcs.ts
// GENERATED CODE -- DO NOT EDIT!
/* tslint:disable */
/* eslint-disable */
import { AdditionalPropertyTarget } from './additional-property-target';
import {EnterpriseEnterpriseSiteVcs} from "./enterprise-enterprise-site-vcs";

/**
 * The top level container
 */
export interface Vcs {

  /**
   * List of virtual cellular services
   */
  vcs?: Array<EnterpriseEnterpriseSiteVcs>;

  [key: string]: AdditionalPropertyTarget | Array<EnterpriseEnterpriseSiteVcs> | undefined;
}